Create link for quiz in leaderboard link back, leaderboard in quiz_detail

+++ quiz/models.py
@@ -127,7 +127,7 @@ class AttemptedQuiz(models.Model):
     quiz_profile = models.ForeignKey(Profile, on_delete=models.CASCADE, related_name='attempts_quiz')
     date_of_birth = models.DateField(auto_now_add=True, blank=True, null=True)
     total_score = models.DecimalField(_('Total Score'), default=0, decimal_places=2, max_digits=10)
-    attempt_number = models.SmallIntegerField(_('Number Attempt'), default=0, null=False)
+    attempt_number = models.PositiveIntegerField(_('Number Attempt'), default=0, null=False)
 
     def get_absolute_url(self):
         return f'/submission-result/{self.pk}/'
